Define GraphQL data model
The data model will support a powerful, intuitive, dynamic, and
deterministic query interface in clients.

While this commit defines a complete, self-contained data model,
the functionality is currently limited to exporting a GraphQL
schema. Actually executing queries is not yet supported because the
PostgreSQL backend is not yet implemented. In addition, the data
model in its current state will suffice for testing and development,
but many more details will need to be added to the ontology, such
as bill history events, legislative chambers and sessions, and
more.