Distinguish between input fields and regular fields.
Regular fields come in many varieties: explicit fields which are
directly represented in the underlying data model, serial fields
which are explicit but whose values are autogenerated on insert,
and inverse-relation fields which do not actually exist on the
type they belong to but are computed via a join based on the presence
of a corresponding field on another type.

Input fields, on the other hand, are always explicitly a part of
the underlying data model, and, as such, are required as input when
inserting new objects.

This distinction allows us to elide the auto-incrementing ID fields
when creating a new object, and let them be populated automatically.
Soon, it will also allow us to skip inverse-relation fields when
describing a new object, since those fields are computed based on
the state of the rest of the database, and specifying them explicitly
wouldn't make sense. In particular, this will help to support plural
fields, which are always inverse relation fields (either the inverse
of a field on another resource, in a has-many relationship, or the
inverse of a field on a join table, in a has-and-belongs-to-many
relationship).