Fix regression in SideMenu styling
In f8e730b203977cffd065f05376b4c69c37a42b60, we added an
`overflow: hidden` style property to elements with a rounded border.
This erroneously applied to the list of selected items in the
multi-select sidebar components, since the `SideMenuItem` component
applied a rounded border style. This meant that, when enough items
were selected to cause a line wrap, the multi-line component
overflowed its parent and was clipped.

In fact, the rounded border style was meant to apply only to buttons
and links, not all side menu items. Removing it allows the selected
items list to resize vertically as the elements wrap across lines.